edit updatejob api, fix coordinate type, restore old create one tag api, fix recommend null job, fix api restore deleted job, fix 400 error message not return

+++ src/entity/address.entity.ts
@@ -32,11 +32,11 @@ export class Address extends Base {
   description: string;
 
   @IsOptional({ groups: [UPDATE, CREATE] })
-  @Column({ type: 'decimal', nullable: true })
+  @Column({ type: 'float', nullable: true })
   latitude: number;
 
   @IsOptional({ groups: [UPDATE, CREATE] })
-  @Column({ type: 'decimal', nullable: true })
+  @Column({ type: 'float', nullable: true })
   longitude: number;
 
   /**
